Changed how "survey summaries" are calculated.
Instead of having a separate dataframe for each of the 4 different kinds
of summaries, there is now one dataframe that can be used to create each
of the different types.

This will be slightly faster (since there is 1 fewer summary to calculate)
but it is also useful to have the single dataframe of all survey summaries
available.